requirements.txt
gym==0.26.2
stable-baselines3==2.0.0
numpy==1.24.3
pygame==2.4.0
pyyaml==6.0
opencv-python==4.8.1.78
matplotlib==3.7.2
torch==2.0.1
shimmy==1.0.0